Rename 'windows' folder to 'browser-windows'
