+++ kiosk/src/js/current_screen_settings_manager.js
@@ -2,11 +2,11 @@
 
 const {LocalStorageManager} = require('./helpers/local_storage_helper');
 const {scheduledTaskManager, isScheduled} = require('./scheduled_task_manager');
-const UserInteractionsManager = require('./services/windows/user_interactions_manager');
+const UserInteractionsManager = require('./services/browser-windows/user_interactions_manager');
 const StorageManager = require('./helpers/storage_manager');
 const SettingsHelper = require('./helpers/settings_helper');
 const DataLoader = require('./services/data/data_loader');
-const WindowInstanceHolder = require('./services/windows/window_instance_holder');
+const WindowInstanceHolder = require('./services/browser-windows/window_instance_holder');
 
 const _ = require('lodash');
