+++ kiosk/src/js/services/browser-windows/user_interactions_manager.js
