+++ kiosk/src/js/services/browser-windows/window_instance_holder.js
