+++ kiosk/src/js/services/browser-windows/windows_helper.js
