+++ kiosk/src/js/services/events/internal-events-listeners/renderer_events_handler.js
@@ -2,8 +2,8 @@ const {ipcMain} = require('electron');
 
 const CurrentScreenSettingsManager = require('../../../current_screen_settings_manager');
 const Logger = require('../../logger/logger');
-const UserInteractionsManager = require('../../windows/user_interactions_manager');
-const WindowsHelper = require('../../windows/windows_helper');
+const UserInteractionsManager = require('../../browser-windows/user_interactions_manager');
+const WindowsHelper = require('../../browser-windows/windows_helper');
 
 /**
  * Should be used after application's 'ready' event was triggered
